refactor existing api to accommodate upcoming version new features
