+++ pom.xml
@@ -9,10 +9,10 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.github.TheDreigon</groupId>
-    <artifactId>JavaCalendarAPI</artifactId>
+    <artifactId>JavaInterviewCalendarAPI</artifactId>
     <version>2.0</version>
-    <name>JavaCalendarAPI</name>
-    <description>Personal Calendar Java project using Spring Boot and REST endpoints</description>
+    <name>JavaInterviewCalendarAPI</name>
+    <description>Personal Interview Calendar Java project using Spring Boot and REST endpoints</description>
     <properties>
         <java.version>19</java.version>
     </properties>
