--- src/main/java/com/github/TheDreigon/JavaCalendarAPI/error/ErrorMessage.java
@@ -1,9 +1,0 @@
-package com.github.TheDreigon.JavaCalendarAPI.error;
-
-/**
- * Error messages to be used throughout the application
- */
-public class ErrorMessage {
-
-    public static final String CALENDAR_NOT_FOUND = "Calendar does not exist";
-}
