--- src/main/java/com/github/TheDreigon/JavaCalendarAPI/exception/CalendarException.java
@@ -1,13 +1,0 @@
-package com.github.TheDreigon.JavaCalendarAPI.exception;
-
-/**
- * A generic calendar exception to be used as base for concrete types of exceptions
- *
- * @see Exception
- */
-public class CalendarException extends Exception {
-
-    public CalendarException(String message) {
-        super(message);
-    }
-}
