+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/controller/InterviewCalendarErrorController.java
@@ -1,4 +1,4 @@
-package com.github.TheDreigon.JavaCalendarAPI.controller;
+package com.github.TheDreigon.JavaInterviewCalendarAPI.controller;
 
 import org.springframework.boot.web.servlet.error.ErrorController;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
  * REST controller responsible for handling error paths
  */
 @RestController
-public class CalendarErrorController implements ErrorController {
+public class InterviewCalendarErrorController implements ErrorController {
     private static final String PATH = "/error";
 
     public String getErrorPath() {
