+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/dto/converter/CalendarDtoToCalendar.java
@@ -1,8 +1,8 @@
-package com.github.TheDreigon.JavaCalendarAPI.dto.converter;
+package com.github.TheDreigon.JavaInterviewCalendarAPI.dto.converter;
 
-import com.github.TheDreigon.JavaCalendarAPI.dto.CalendarDto;
-import com.github.TheDreigon.JavaCalendarAPI.persistence.model.Calendar;
-import com.github.TheDreigon.JavaCalendarAPI.service.CalendarService;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.CalendarDto;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.Calendar;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.service.CalendarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
