+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/error/ErrorMessage.java
@@ -1,0 +1,11 @@
+package com.github.TheDreigon.JavaInterviewCalendarAPI.error;
+
+/**
+ * Error messages to be used throughout the application
+ */
+public class ErrorMessage {
+
+    public static final String CALENDAR_NOT_FOUND = "Calendar not found";
+    public static final String CANDIDATE_NOT_FOUND = "Candidate not found";
+    public static final String INTERVIEWER_NOT_FOUND = "Interviewer not found";
+}
