+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/exception/JavaInterviewCalendarAPIException.java
@@ -1,0 +1,13 @@
+package com.github.TheDreigon.JavaInterviewCalendarAPI.exception;
+
+/**
+ * A generic JavaInterviewCalendarAPI exception to be used as base for concrete types of exceptions
+ *
+ * @see Exception
+ */
+public class JavaInterviewCalendarAPIException extends Exception {
+
+    public JavaInterviewCalendarAPIException(String message) {
+        super(message);
+    }
+}
