+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/persistence/model/AbstractModel.java
@@ -1,4 +1,4 @@
-package com.github.TheDreigon.JavaCalendarAPI.persistence.model;
+package com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model;
 
 import jakarta.persistence.*;
 import lombok.Getter;
