+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 logging.level.web=TRACE
 logging.level.org.springframework.web=TRACE
 
-spring.datasource.url=jdbc:h2:mem:calendarDb
+spring.datasource.url=jdbc:h2:mem:interviewCalendarDb
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
